fix chat
