+++ client/package.json
@@ -22,6 +22,7 @@
         "react-dom": "^18.2.0",
         "react-hook-form": "^7.33.1",
         "react-icons": "^4.4.0",
+        "react-input-emoji": "^4.1.0",
         "react-paypal-button-v2": "^2.6.3",
         "react-redux": "^8.0.2",
         "react-router-bootstrap": "^0.26.2",
