--- client/src/chat/css/MessageForm.css
@@ -1,67 +1,0 @@
-.messages-output {
-    /* background-image: url(https://img.freepik.com/free-vector/abstract-watercolor-pastel-background_87374-137.jpg?w=360);
-    background-position: center;
-    background-size: cover; */
-    background-color: white;
-    height: 80vh;
-    border: 1px solid lightgray;
-    overflow-y: scroll;
-    margin-bottom: 20px;
-}
-
-.message-inner {
-    /* margin-left: 20px; */
-    margin-bottom: 10px;
-    padding: 10px;
-    min-width: 200px;
-    max-width: 90%;
-    text-align: left;
-    min-height: 80px;
-    /* font: 400 1em, sans-serif; */
-    border-radius: 10px;
-    background-color: #d1e7dd;
-    overflow: hidden;
-}
-
-.incoming-message .message-inner {
-    background-color: #ffdab9;
-}
-
-.message {
-    display: flex;
-    justify-content: flex-start;
-    margin-left: 5px;
-    margin-right: 20px;
-}
-
-.message-content {
-    margin-top: 5px;
-    font: 400 1em, sans-serif;
-    font-size: 1.1em
-}
-
-.incoming-message {
-    display: flex;
-    justify-content: flex-end;
-    margin-right: 5px;
-    margin-left: 20px;
-}
-
-.message-timetamp-left {
-    font-size: 0.85em;
-    font-weight: 300;
-    margin-top: 10px;
-    margin-bottom: 0px;
-}
-
-.message-sender {
-    margin-bottom: 5px;
-    font-weight: bold;
-}
-
-.message-date-indicator {
-    width: 150px;
-    margin: 15px auto;
-    /* padding: 100px; */
-    border-radius: 10px;
-}
