--- client/src/chat/css/chat.css
@@ -1,6 +1,0 @@
-.chat__bg {
-    background-image: url(https://img.freepik.com/free-vector/shiny-white-gray-background-with-wavy-lines_1017-25101.jpg?w=2000);
-    background-position: center;
-    background-size: cover;
-    min-height: 92vh;
-}
